Factored out chunks into separate module
